story/CCMSPUI-845
claim errors tab
