+++ laa-submit-a-bulk-claim-ui/src/main/resources/messages.properties
@@ -67,6 +67,7 @@ bulkClaimImportSummary.errorsFoundModal.message=Resolve the errors and upload th
 bulkClaimImportSummary.claimErrors.caption=Claim errors
 viewSubmission.heading=Submission details
 viewSubmission.matterStarts.subHeading=Matter starts
+viewSubmission.claimErrors.subHeading=Claim errors
 viewSubmission.claims.subHeading=Claims
 viewSubmission.claims.submissionCosts.caption=Submission costs summary
 viewSubmission.claims.submissionCosts.tableHeader.name=Name
