+++ laa-submit-a-bulk-claim-ui/src/main/resources/templates/fragments/claim-errors.html
@@ -1,0 +1,43 @@
+<table class="govuk-table"
+       th:if="${errorDetails.containsErrors()}"
+       data-module="moj-sortable-table"
+       th:fragment="claimErrors(errorDetails)">
+  <caption class="govuk-table__caption govuk-table__caption--m"
+           th:text="#{bulkClaimImportSummary.claimErrors.caption}"></caption>
+
+  <thead class="govuk-table__head">
+  <tr class="govuk-table__row">
+    <th scope="col" class="govuk-table__header" aria-sort="none"
+        th:text="#{submission.claim.properties.ufn}"></th>
+    <th scope="col" class="govuk-table__header" aria-sort="none"
+        th:text="#{submission.claim.properties.ucn}"></th>
+    <th scope="col" class="govuk-table__header" aria-sort="none"
+        th:text="#{submission.claim.properties.client}"></th>
+    <th scope="col" class="govuk-table__header" aria-sort="none"
+        th:text="#{submission.claim.properties.errorDescription}"></th>
+  </tr>
+  </thead>
+
+  <tbody class="govuk-table__body">
+  <tr class="govuk-table__row"
+      th:each="error, idx : ${errorDetails.claimErrors()}"
+      th:id="|claim-error-${idx.index}|">
+    <td class="govuk-table__cell"
+        th:id="|error-ufn-${idx.index}|"
+        th:data-sort-value="${error.ufn()}"
+        th:text="${error.ufn()}"></td>
+    <td class="govuk-table__cell"
+        th:id="|error-ucn-${idx.index}|"
+        th:data-sort-value="${error.ucn()}"
+        th:text="${error.ucn()}"></td>
+    <td class="govuk-table__cell"
+        th:id="|error-client-${idx.index}|"
+        th:data-sort-value="${error.client()}"
+        th:text="${error.client()}"></td>
+    <td class="govuk-table__cell"
+        th:id="|error-description-${idx.index}|"
+        th:data-sort-value="${error.message()}"
+        th:text="${error.message()}"></td>
+  </tr>
+  </tbody>
+</table>
