+++ laa-submit-a-bulk-claim-ui/src/test/java/uk/gov/justice/laa/bulkclaim/controller/SubmissionDetailControllerTest.java
@@ -26,6 +26,7 @@ import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.assertj.MockMvcTester;
 import reactor.core.publisher.Mono;
 import uk.gov.justice.laa.bulkclaim.builder.SubmissionClaimDetailsBuilder;
+import uk.gov.justice.laa.bulkclaim.builder.SubmissionClaimErrorsBuilder;
 import uk.gov.justice.laa.bulkclaim.builder.SubmissionMatterStartsDetailsBuilder;
 import uk.gov.justice.laa.bulkclaim.builder.SubmissionSummaryBuilder;
 import uk.gov.justice.laa.bulkclaim.client.DataClaimsRestClient;
@@ -49,6 +50,7 @@ class SubmissionDetailControllerTest {
   @MockitoBean private SubmissionClaimDetailsBuilder submissionClaimDetailsBuilder;
   @MockitoBean private SubmissionMatterStartsDetailsBuilder submissionMatterStartsDetailsBuilder;
   @MockitoBean private DataClaimsRestClient dataClaimsRestClient;
+  @MockitoBean private SubmissionClaimErrorsBuilder submissionClaimErrorsBuilder;
 
   @Nested
   @DisplayName("GET: /submission/{submissionId}")
